Refactor `rse-plugin` to be a bit more of a "core" crate
Most crates in the workspace should be "core" crates - they only contain core functionality without *too many* opinionated abstractions.

Historically, the whole structure of the workspace was *drastically* different. `rse-plugin` was the "main" crate that contained all of the high-level functionality. However, further development proved that a more opinionated "standard" library is needed, which `rse-std` (proper name pending) aims to be. So, `rse-plugin` won't re-export crates that it doesn't need to.

It didn't really make sense for `InterfaceFactories<'_>` to exist in `rse-game-interfaces`, since the plugins that are inside are plugin-specific, so this type was moved to `rse-plugin`. The lifetime also didn't really make sense, as the factory functions themselves will definitely outlive the plugin.